Add files via upload
This R script relates to the manuscript "Juvenile idiopathic arthritis polygenic risk scores are associated with cardiovascular phenotypes in early adulthood". It details the ALSPAC variables used in this analysis, details of data cleaning and variable recoding, but excludes the variable extraction from the ALSPAC data catalogue as this differs between internal/external data users.